Move card components to appropriate directory
